#6 Pass hosting bucket name and frontend source path to the construct

+++ provisioning/lib/hosting-bucket.ts
@@ -2,21 +2,31 @@ import s3 = require("@aws-cdk/aws-s3");
 import {Construct, RemovalPolicy} from "@aws-cdk/core";
 import {BucketDeployment, Source} from "@aws-cdk/aws-s3-deployment";
 
-export class HostingBucket extends Construct {
-    constructor(scope: Construct, id: string) {
-        super(scope, id);
-        const deploymentBucket = new s3.Bucket(this, 'deployment-bucket', {
-            bucketName: 'z10020-dwp-hack-the-north',
-            websiteIndexDocument: 'index.html',
-            removalPolicy: RemovalPolicy.DESTROY, // bucket must be empty before performing 'cdk destroy'
-        });
-
-        new BucketDeployment(this, 'deploy-website', {
-            sources: [Source.asset('../dwp-hack-the-north-ui/build')],
-            destinationBucket: deploymentBucket
-        });
-
-        deploymentBucket.grantPublicAccess()
-    }
+export interface HostingBucketProps {
+    bucketName: string
+    sourcePath: string
 }
 
+export class HostingBucket extends Construct {
+    constructor(scope: Construct, id: string, props: HostingBucketProps = {bucketName: '', sourcePath: ''}) {
+        super(scope, id);
+
+        if ( (props.bucketName != null && props.bucketName != "") && (props.sourcePath != null && props.sourcePath != "")) {
+            const deploymentBucket = new s3.Bucket(this, 'deployment-bucket', {
+                bucketName: props.bucketName,
+                websiteIndexDocument: 'index.html',
+                removalPolicy: RemovalPolicy.DESTROY, // bucket must be empty before performing 'cdk destroy'
+            });
+
+            new BucketDeployment(this, 'deploy-website', {
+                sources: [Source.asset(props.sourcePath)],
+                destinationBucket: deploymentBucket
+            });
+
+            deploymentBucket.grantPublicAccess()
+        } else {
+            console.error('ERROR: Bucket name and source path for the frontend must be provided')
+        }
+
+    }
+}
